breaking change: WIA not required anymore in openid4vp - moved in the top folder
