+++ pyeudiw/tests/wallet_instance_attestation/__init__.py
@@ -1,0 +1,1 @@
+# Wallet instance attestation tests
